MSI-BruiseNet/requirements.txt
torch
torchvision
numpy
scipy
matplotlib
seaborn
scikit-learn
opencv-python
albumentations
tensorboard
pyyaml
tqdm
thop
